% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_compare_country.R
\name{compare_country}
\alias{compare_country}
\title{Compare the Existing Country Names with the Determined Country Names.}
\usage{
compare_country()
}
\description{
Compare the given country names with the names assigned by the function
\code{\link{get_country}} and also by the functions \code{\link{get_sea}} and
\code{\link{get_country_sea}}, if applicable.
}
\details{
\itemize{
\item Input:
\enumerate{
\item The file 'geographic_names.txt', created by
\code{\link{order_data}}, which contains the names assigned by the function
\code{\link{get_country}} and also by the functions \code{\link{get_sea}} and
\code{\link{get_country_sea}}, if applicable.
\item The file used as input of \code{\link{get_country}} with the given
names to compare with the ones determinated by the functions.
}
\item The function does the following:
\enumerate{
\item Writes a file with all the points that have a different name;
\item Writes a file with the points that have a different name for the same
country (both names correct but different);
\item Writes a file with the points that have a country name assigned to the
given name and have the  Exclusive Economic Zone name of that country
assigned to the determinated name. Probably this points are very close to
the coastline and the associated positioning error causes them to be
effectively located at sea. It's a tolerance issue;
\item Writes a file whose points have unrelated names. This indicates that
they are probably incorrectly located, consisting in errors that must be
corrected;
\item Writes a processing report.
}
\item Output:
\enumerate{
\item A .txt file named 'different_names.txt';
\item A .txt file named 'err_just_name.txt';
\item A .txt file named 'err_just_tol.txt';
\item A .txt file named 'err_location.txt';
\item A .txt file named 'report_compare_country.txt'.
}
}
}
