% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_data.R
\docType{data}
\name{eraclim_uao_fp}
\alias{eraclim_uao_fp}
\title{ERACLIM Upper-air sations}
\format{A data frame with 6444 rows and 3 variables:
\describe{
  \item{lon}{longitude in decimal degrees with three decimal places, that
  should be in the range (-180 to 180) but has out of bounds values and
  missing values}
  \item{lat}{latitude in decimal degrees with three decimal places, which
  has some missing values}
  \item{country_gv}{country name given in the inventory, which is written in
  different languages and has non ASCII characters}
}}
\source{
\url{http://eraclim-global-registry.rd.ciencias.ulisboa.pt/upperairdata.php}
}
\usage{
eraclim_uao_fp
}
\description{
A dataset of coordinate pairs with the longitude in the range [-180, +180]
and respective country, extracted from the \strong{ERACLIM Upper-air
Observations - Fixed Platforms} inventory.
}
\keyword{datasets}
