% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country}
\alias{get_country}
\title{Determines the country names for coordinate points}
\usage{
get_country()
}
\description{
Given a pair of geographic coordinates in the continent, determines the
country name.
}
\details{
Input: A text file with the coordinates in decimal degrees and a
  shapefile with the country names.

  The .txt file has to be in the following format: with the longitude in the
  first column and the latitude in the second column, separated by tabs.The
  header of the file must be 'longitude' and 'latitude', respectively and
  longitude must be in the range [-180, +180].

  If exists <coords2getf.txt> created by \code{\link{get_lon180}},
  automatically loads and reads the coordinates file. If wasn't necessary to
  use \code{\link{get_lon180}}, because the longitude was already in the
  range [-180, +180], opens a window, for choosing the input file that must
  be in the working directory.

  Creates an 'id' with the row number, which consists of an identifier to the
  coordinates: the 'id' is necessary to display the coordinates in the
  initial order at the end of the process.

  Output: writes the country names in <countries.txt>; writes the missing
  name points (if they exist) in <countries_na.txt>, which is the input of
  \code{\link{get_marine}}; writes a processing report in
  <report_get_country.txt>.
}
\examples{
E.g. the Portuguese territory
If the point is in Continental Portugal
attributes the name 'Portugal'
If the point is in the Portuguese
islands Azores or Madeira
which are the two Autonomous Regions
of the Portuguese sovereign
attributes 'Azores' and 'Madeira'
respectively.

}
\references{
1:10m Cultural Vectors: Admin 0 - Countries. Made with Natural
  Earth. Free vector and raster map data @ naturalearthdata.com. Consulted
  on 2017-11-15.
}
\seealso{
Requires \code{\link[rgdal]{readOGR}},
  \code{\link[sp]{SpatialPoints}} and \code{\link[sp]{over}} to determine the
  country names.
}
