% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_get_country_shoreline.R
\name{get_country_shoreline}
\alias{get_country_shoreline}
\title{Determines Country Names for Coordinate Points Located in the Shoreline.}
\usage{
get_country_shoreline(icoords, tol)
}
\arguments{
\item{icoords}{data frame with three columns: \strong{id | lon | lat} where:
'id' is the row identifier for the coordinates in the original list of
coordinates, 'lon' is the longitude in the range [-180, +180] and 'lat' is
the latitude. Both coordinates are in decimal degrees.}

\item{tol}{is the tolerance in meters. By default, tol = 500 meters.}
}
\description{
Determines country names for a list of coordinate points which are supposed
to be located on land but without country names assigned by the function
\code{\link{get_country}} because probably they are very close to the
shoreline and have lack of precision. Due to this fact when the user runs
\code{\link{get_sea}} over that points they get ocean/sea names assigned. The
present function uses a SpatialPolygonsDataFrame with 50 m precision and a
buffer zone around the coordinate points. The buffer represents the tolerance
and is given by the user. If the tolerance is 500 m, for example, and the
point gets a country name, means that it belongs to that country with an
error until 500 m towards the sea.
}
\details{
\strong{Input:}
\itemize{
\item The output of \code{\link{get_country}}, i.e., a data frame with three
columns: \strong{id | lon | lat}.
}
\strong{Output:}
\itemize{
\item Eventually, a text file with the points for which a country name was
determined. That file has the following header: \strong{id | lon | lat |
country | sovereignt | adm0_a3 | name_de | name_es | name_fr | name_pt}.
\item Eventually, a text file with the points missing country names.
\item A .RData file with the country names data frame - 'countries_sh' -, the
missing country names data frame - 'miss_countries_sh' - or the both.
'miss_countries_sh' is the input of \code{\link{get_sea}} to determine the
remaining names.
}
}
\examples{
\dontrun{
##
## First run
get_lon180(coords_lon360)
## Or
test_geocoord(coords_sample)
## Then run sequentially
get_country(icoords)
get_country_shoreline(icoords, tol)
}

}
\references{
Made with Natural Earth. Free vector and raster map data @
naturalearthdata.com.
}
\seealso{
Requires \code{\link[raster]{buffer}}.
}
