% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_get_lon180.R
\name{get_lon180}
\alias{get_lon180}
\title{Tests the Geographic Coordinates and Transforms the Longitude from (0, 360)
to (-180, +180).}
\usage{
## If there are the coordinates data frame
get_lon180(coords)
##
## If there are the coordinates text file
get_lon180(coords = NULL)
}
\arguments{
\item{coords}{data frame with the geographic coordinates in decimal degrees,
the longitude in the first column, the latitude in the second column and
the column names: \strong{lon | lat} (other columns can exist, however are
unnecessary for this function).}
}
\description{
Given a list of geographic coordinates, first tests the coordinates for out
of bounds values and missing values. Then transforms the longitude from (0,
360) to (-180, +180) degrees, considering 6 significative digits.
}
\details{
\strong{Input:}
\itemize{
\item A \strong{text file} without header with the geographic coordinates in
decimal degrees: the longitude in the first column and the latitude in the
second column, separated by tabs (other columns could exist, however are
unnecessary for this function). Missing values must be codified as 'NA' in
all fields.
\item Or a \strong{data frame} with that same format and the column names:
\strong{lon | lat}.
}
\strong{Output:}
\itemize{
\item A text file named 'coords_lon180' with three columns id | lon | lat,
where: 'id' is the row identifier for the coordinates in the original list of
coordinates, 'lon' is the longitude in the range (-180, +180) and 'lat' is
the latitude. Both coordinates are in decimal degrees.
\item If there are errors, the function writes a text file with the
coordinate pairs containing at least one coordinate out of bounds.
\item If there are missing coordinates, the function writes a text file with
the coordinate pairs containing at least one missing coordinate.
\item A .RData file with the output data frame(s) that has/have the column
names: \strong{id | lon | lat}. The data frame \strong{'coords_lon180'} can
be used as input parameter of \code{\link{get_country}} to determine the
country names. Eventually is created the data frame \strong{'excl_coords'}
with the erroneous and/or missing coordinates.
}
}
\examples{
\dontrun{
get_lon180(coords = ispd)
}

}
