% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_get_marine.R
\name{get_marine}
\alias{get_marine}
\title{Determines the marine regions names for coordinate points}
\usage{
get_marine()
}
\description{
Given a pair of geographic coordinates in a marine region, determines the
marine region name.
}
\details{
Input: A text file with the coordinates in decimal degrees and a
  shapefile with the marine regions names (resulting from the union of the
  Exclusive Economic Zones (EEZ) with the International Hydrographic
  Organization (IHO) Seas.

  The .txt file has to be in the following format: with the longitude in the
  first column and the latitude in the second column, separated by tabs.The
  header of the file must be 'longitude' and 'latitude', respectively and
  longitude must be in the range [-180, +180].

  If exists <countries_na.txt> created by \code{\link{get_country}},
  automatically loads and reads the coordinates file. If wasn't necessary to
  use \code{\link{get_country}}, because all the points are in a marine
  region, opens a window for choosing the input file that must be in the
  working directory.

  Preserves the 'id' assigned to the coordinates in \code{\link{get_country}}
  or, for the case of \code{\link{get_country}} wasn't necessary, creates an
  'id' with the row number, which consists of an identifier to the
  coordinates: the 'id' is necessary to display the coordinates in the
  initial order at the end of the process.

  Output: writes the marine regions names in <marine_regions.txt>; writes the
  missing name points (if they exist) in <marine_regions_na.txt>, which is
  the input of \code{\link{get_country_sea}}; writes a processing report in
  <report_get_marine.txt>.
}
\examples{
E.g. the Portuguese territory
If the point is in the
Portuguese Continental Shelf
attributes the name
'Portuguese Part of the
Atlantic Ocean'.
If the point isnt part of a
continental shelf
attributes the name of the
sea/ocean
e.g. 'North Atlantic Ocean'.

}
\references{
VLIZ (2012). Intersect of IHO Sea Areas and Exclusive Economic
  Zones (version 2). Available online at http://www.marineregions.org/.
  Consulted on 2017-11-15.
}
\seealso{
Requires \code{\link[rgdal]{readOGR}},
  \code{\link[sp]{SpatialPoints}} and \code{\link[sp]{over}} to determine the
  country names.
}
