% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_get_sea.R
\name{get_sea}
\alias{get_sea}
\title{Determines Sea Names for Points Located in a Marine Area.}
\usage{
get_sea(icoords)
}
\arguments{
\item{icoords}{data frame with three columns: \strong{id | lon | lat} where:
'id' is the row identifier for the coordinates in the original list of
coordinates, 'lon' is the longitude in the range (-180, +180) and 'lat' is
the latitude. Both coordinates are in decimal degrees.}
}
\description{
Given a list of geographic coordinates in marine areas of the world,
determines the geographical name of the sea, ocean, bay, gulf, fjord, etc.
were each one of the points is located. Uses a SpatialPolygonsDataFrame with
10 m precision containing the polygons and sea names, which covers most of
the marine territories on Earth.
}
\details{
\strong{Input:}
\itemize{
\item The output of \code{\link{get_country}} - \strong{'miss_countries'} -,
or the output of \code{\link{get_country_shoreline}}, -
\strong{'miss_countries_sh'} -, i.e., a data frame with three columns:
\strong{id | lon | lat}.
}
\strong{Output:}
\itemize{
\item A text file with the points for which a sea name was determined, if
they exist. That file has the following header: \strong{id | lon | lat |
sea}.
\item A text file with the missing sea names, if they exist.
\item A .RData file with the output data frame(s). The data frame
\strong{'seas'} has the header \strong{id | lon | lat | sea}. The eventual
data frame \strong{'miss_seas'} has the header \strong{id | lon | lat}.
}
}
\examples{
\dontrun{
## First run
get_lon180(coords = ispd)
## Or
test_geocoord(coords = eraclim_uao_fp)
## Then run sequentially
get_country(icoords = coords_ok)
get_country_shoreline(icoords = miss_countries, tol)
get_sea(icoords = miss_countries_sh)
}

}
\references{
Made with Natural Earth. Free vector and raster map data @
naturalearthdata.com.
}
