% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_data.R
\docType{data}
\name{ispd}
\alias{ispd}
\title{ISPD stations}
\format{A data frame with 83814 rows and 2 variables:
\describe{
  \item{lon}{longitude in decimal degrees (0 to 360) with two decimal places}
  \item{lat}{latitude in decimal degrees (-90 to +83.65) with two decimal
  places}
}}
\usage{
ispd
}
\description{
A dataset of coordinate pairs with the longitude in the range (0 to 360)
degrees, extracted from the list of meteorological stations belonging to
\emph{ISPD - International Surface Pressure Databank}.
}
\references{
\url{http://reanalyses.org/observations/international-surface-pressure-databank}
}
\keyword{datasets}
