% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-rgeoname.R
\docType{package}
\name{rgeoname}
\alias{rgeoname}
\alias{rgeoname-package}
\title{rgeoname: A package for the quality control of meteorological stations location}
\description{
rgeoname: A package for the quality control of meteorological stations location
}
\section{rgeoname functions}{

get_lon180, get_country, get_marine, get_country_sea and order_data.
}

