% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-stlocationqc.R
\docType{package}
\name{stlocationqc}
\alias{stlocationqc}
\alias{stlocationqc-package}
\title{stlocationqc: C3S Quality Control of Meteorological Stations Location}
\description{
stlocationqc: C3S Quality Control of Meteorological Stations Location
}
\section{stlocationqc functions}{

\code{\link{countries_polys}}, \code{\link{seas_polys}},
\code{\link{get_lon180}}, \code{\link{test_geocoord}},
\code{\link{get_country}}, \code{\link{get_country_shoreline}},
\code{\link{get_sea}}, \code{\link{order_data}} and
\code{\link{compare_country}}.
}

