% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_test_geocoord.R
\name{test_geocoord}
\alias{test_geocoord}
\title{Tests the Geographic Coordinates.}
\usage{
## If there are the coordinates data frame
test_geocoord(coords)
##
## If there are the coordinates text file
test_geocoord(coords = NULL)
}
\arguments{
\item{coords}{data frame with the geographic coordinates in decimal degrees,
the longitude in the first column, the latitude in the second column and
the column names: \strong{lon | lat} (other columns could exist, however
are unnecessary for this function).}
}
\description{
Given a list of geographic coordinates, tests if the longitude belongs to
(-180, +180) degrees and if the latitude belongs to (-90, +90) degrees. Also
tests the coordinates for missing values.
}
\details{
\strong{Input:}
\itemize{
\item A \strong{text file} without header with the geographic coordinates in
decimal degrees: the longitude in the first column and the latitude in the
second column, separated by tabs (other columns could exist, however are
unnecessary for this function). Missing values must be codified as 'NA' in
all fields.
\item Or a \strong{data frame} with that same format and the column names:
\strong{lon | lat}.
}
\strong{Output:}
\itemize{
\item A text file named 'coords_ok' with three columns id | lon | lat, where:
'id' is the row identifier for the coordinates in the original list of
coordinates, 'lon' is the longitude in the range (-180, +180) and 'lat' is
the latitude. Both coordinates are in decimal degrees.
\item If there are errors, the function writes a text file with the
coordinate pairs containing at least one coordinate out of bounds.
\item If there are missing coordinates, the function writes a text file with
the coordinate pairs containing at least one missing coordinate.
\item A .RData file  with the output data frame(s) that has/have the column
names: \strong{id | lon | lat}. The data frame \strong{'coords_ok'} can be
used as input parameter of \code{\link{get_country}} to determine the country
names. Eventually is created the data frame \strong{'excl_coords'} with the
erroneous and/or missing coordinates.
}
}
\examples{
\dontrun{
test_geocoord(coords = eraclim_uao_fp)
}

}
