% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_compare_country.R
\name{compare_country}
\alias{compare_country}
\title{Compares the Given Country Names with the Determined Country Names.}
\usage{
compare_country(countries_gv = NULL, countries, countries_sh = NULL, seas =
NULL, miss_seas = NULL, excl_coords = NULL)
}
\arguments{
\item{countries_gv}{data frame with the coordinates and the given country
names \strong{lon | lat | country_gv}.}

\item{countries}{data frame output of \code{\link{get_country}}.}

\item{countries_sh}{data frame output of \code{\link{get_country_shoreline}}.}

\item{seas}{data frame output of \code{\link{get_sea}}.}

\item{miss_seas}{data frame output of \code{\link{get_sea}}.}

\item{excl_coords}{data frame output of \code{\link{get_lon180}} or
\code{\link{test_geocoord}}.}
}
\description{
Compares the given country names with the names assigned by the function
\code{\link{get_country}}, \code{\link{get_country_shoreline}} or
\code{\link{get_sea}}.
}
\details{
\strong{Input:}
\itemize{
\item A data frame with the coordinates and the given country names or a text
file without header, the longitude in the first column, the latitude in the
second column, both in decimal degrees. The given country name goes in the
third column. The columns are separated by tabs and missing values must be
codified as 'NA' in all fields.
\item The data frames with names assigned to the coordinates, i.e., the
output of \code{\link{get_country}}, \code{\link{get_country_shoreline}} or
\code{\link{get_sea}}. If exists 'miss_seas' or 'excl_coords' (output of
\code{\link{test_geocoord}} / \code{\link{get_lon180}}), they should be
considered.
}
\strong{Output:}
\itemize{
\item Several text files, until the maximum of 9, divided by types of
differences in the country name and the following header: \strong{country_gv
|id | lon | lat | country | sovereignt | adm0_a3 | name_de | name_es |
name_fr | name_pt}.
}
}
\examples{
\dontrun{
##
## First run
test_geocoord(coords = eraclim_uao_fp)
## Then run sequentially until all points have names assigned
get_country(icoords)
get_country_shoreline(icoords, tol)
get_sea(icoords)
## And finally
compare_country(countries_gv = NULL, countries, countries_sh = NULL, seas =
NULL, miss_seas = NULL, excl_coords = NULL)
}

}
