% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_compare_country.R
\name{compare_country}
\alias{compare_country}
\title{Compares the Given Country Names with the Determined Country Names.}
\usage{
compare_country(countries_gv = NULL, countries, countries_sh = NULL, seas =
NULL, miss_seas = NULL, excl_coords = NULL)
}
\arguments{
\item{countries_gv}{data frame with the coordinates and the given country
names \strong{lon | lat | country_gv}.}

\item{countries}{data frame output of \code{\link{get_country}}.}

\item{countries_sh}{data frame output of \code{\link{get_country_shoreline}}.}

\item{seas}{data frame output of \code{\link{get_sea}}.}

\item{miss_seas}{data frame output of \code{\link{get_sea}}.}

\item{excl_coords}{data frame output of \code{\link{get_lon180}} or
\code{\link{test_geocoord}}.}
}
\description{
Compares the given country names with the names assigned by the function
\code{\link{get_country}}, \code{\link{get_country_shoreline}} or
\code{\link{get_sea}}.
}
\details{
\strong{Input:}
\itemize{
\item A data frame with the coordinates and the given country names or a text
file without header, the longitude in the first column, the latitude in the
second column, both in decimal degrees. The given country name goes in the
third column. The columns are separated by tabs and missing values must be
codified as 'NA' in all fields.
\item The output of \code{\link{get_country}} - \strong{'countries'} -,
\code{\link{get_country_shoreline}} - \strong{'countries_sh'} -,
\code{\link{get_sea}} - \strong{'seas'} and \strong{'miss_seas'} - and
\code{\link{get_lon180}} or \code{\link{test_geocoord}} -
\strong{'excl_coords'}. However, only the 'countries' data frame is
compulsory because all those data frames may not exist.
}
\strong{Output:}
\itemize{
\item Several text files, until the maximum of 9, which have the following
header: \strong{country_gv |id | lon | lat | country | sovereignt | adm0_a3 |
name_de | name_es | name_fr | name_pt}. Those files split the list of points
by types of differences between the given and the determinated country name:
\itemize{
\item Missing given or determined country name
\item Equal country name
\item Given country name equals to sovereignty
\item Given country name in another language
\item Given country name in upper case
\item Given country name with some equal words - English or another language
\item Set of 3 letters equals at the beginning or at the end of the words
\item Points that fall into the sea
\item Different country name
}
}
}
\examples{
\dontrun{
##
## First run
test_geocoord(coords = eraclim_uao_fp)
## Then run sequentially until all points have names assigned
get_country(icoords = coords_ok)
get_country_shoreline(icoords = miss_countries, tol)
get_sea(icoords = miss_countries_sh)
## And finally
compare_country(countries_gv = eraclim_uao_fp, countries = countries,
countries_sh = countries_sh, seas = seas, miss_seas = miss_seas, excl_coords
= excl_coords)
}

}
