% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_countries_polys.R
\name{countries_polys}
\alias{countries_polys}
\title{Downloads Spatial Data with the World Countries Polygons and Saves them as R
Objects.}
\usage{
countries_polys()
}
\description{
Downloads from Natural Earth two shapefiles - 10 m and 50 m precision - with
the world countries polygons and transforms them into two
SpatialPolygonsDataFrame that are saved in a folder called 'polys', inside
the user's working directory. The data that they contain is necessary for the
functions \code{\link{get_country}} and \code{\link{get_country_shoreline}}.
}
\details{
The two SpatialPolygonsDataFrame are saved in the working directory, inside
the 'polys' folder as .rda files and they are loaded from there when running
the functions \code{\link{get_country}} and
\code{\link{get_country_shoreline}}.

\strong{Shapefiles:}
\itemize{
\item http://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_countries.zip
\item http://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_countries.zip
}
\strong{Shapefiles metadata:}
\itemize{
\item http://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-0-countries/
\item http://www.naturalearthdata.com/downloads/50m-cultural-vectors/50m-admin-0-countries-2/
}
}
\references{
Made with Natural Earth. Free vector and raster map data @
naturalearthdata.com.
}
