% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_get_country.R
\name{get_country}
\alias{get_country}
\title{Determines Country Names for Coordinate Points Located in the Continent.}
\usage{
get_country(icoords)
}
\arguments{
\item{icoords}{data frame with three columns: \strong{id | lon | lat} where:
'id' is the row identifier for the coordinates in the original list of
coordinates, 'lon' is the longitude in the range [-180, +180] and 'lat' is
the latitude. Both coordinates are in decimal degrees.}
}
\description{
Given a list of geographic coordinates in the continent, determines the
country name for each coordinate point. First uses a SpatialPolygonsDataFrame
with 10 m precision to calculate the country names. Then if there are still
points without a name assigned, uses a SpatialPolygonsDataFrame with 50 m
precision, thus trying to calculate the name for less accurate points.
}
\details{
\strong{Input:}
\itemize{
\item A data frame with three columns \strong{id | lon | lat}, i.e., the
output of \code{\link{test_geocoord}} - 'coords_ok' -, or the output of
\code{\link{get_lon180}} - 'coords_lon180'. To use the present function it is
necessary that the coordinates do not have any errors or missing values. This
requires previous running of \code{\link{test_geocoord}} or
\code{\link{get_lon180}} that in addition to testing the coordinates, creates
the 'id' which is the row identifier for the coordinates in the original list
of coordinates. After running sequentially \code{\link{get_country}},
\code{\link{get_country_shoreline}} and \code{\link{get_sea}} to assign the
geographic names, the 'id' is necessary to display the coordinates in the
initial order at the end of the process (which consists of a sucessive
elimination of missing names), with or without all the names assigned.
}
\strong{Output:}
\itemize{
\item A text file with all the points for which a country name was
determined. That file has the following header: \strong{id | lon | lat |
country | sovereignt | adm0_a3 | name_de | name_es | name_fr | name_pt}.
\item A text file with the missing country names, if they exist.
\item A .RData file  with the output data frame(s). The data frame
'countries' as the header \strong{id | lon | lat | country | sovereignt |
iso3 | subregion | continent}. The data frame 'miss_countries' as the header
\strong{id | lon | lat} and is the input of
\code{\link{get_country_shoreline}} to determine the missing country names or
the input of \code{\link{get_sea}}, if the user wants to verify if the still
unamed points fall into to the sea and calculate the sea name were they are
located.
}
}
\examples{
\dontrun{
##
## Example:
## A sample of coordinate pairs with the longitude in the range [-180, +180],
## extracted from the list of meteorological stations belonging to
## ISPD - International Surface Pressure Databank, ERACLIM upper-air
## inventory and MEDARE - MEditerranean climate DAta REscue
##
## First run
get_lon180(coords_lon360)
## Or
test_geocoord(coords_sample)
## Then run
get_country(icoords)
}

}
\references{
Made with Natural Earth. Free vector and raster map data @
naturalearthdata.com.
}
