% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_country_sea.R
\name{get_country_sea}
\alias{get_country_sea}
\title{Determines the marine regions names for other points}
\usage{
get_country_sea()
}
\description{
Determines the  country names for points in the boundaries of continent with
Exclusive Economic Zones (EEZ) and the sea/ocean/EEZ names for other points
in a marine region.
}
\details{
Input: The file <marine_regions_na.txt> created by
  \code{\link{get_marine}} which contains the missing name points that left.
  Those points are located in the boundaries of continent with marine regions
  and in marine regions not covered by the polygons of the shapefile used by
  \code{\link{get_marine}}; Three shapefiles: the first with country names
  (resulting from the union of EEZ with respective countries), the second
  with the IHO Sea names and the third with the EEZ names.

  Preserves the 'id' assigned to the coordinates in \code{\link{get_country}}
  or in \code{\link{get_marine}}.

  Output: writes the countries and marine regions names in
  <countries_seas.txt>; writes the missing name points (if they exist) in
  <countries_seas_na.txt>; writes a processing report in
  <report_get_country_sea.txt>. The points without a name that eventualy can
  occur after running this function, are integrated in <countries_seas.txt>
  with de designation "NO_NAME".
}
\references{
Flanders Marine Institute (2016). Maritime Boundaries
  Geodatabase, version 1. Available online at http://www.marineregions.org/.
  Consulted on 2017-11-15.

  VLIZ (2014). Union of the ESRI Country shapefile and the Exclusive Economic
  Zones (version 2). Available online at http://www.marineregions.org/.
  Consulted on 2017-11-15.

  VLIZ (2017). IHO Sea Areas, version 2. Available online at
  http://www.marineregions.org/. Consulted on 2017-10-10.
}
\seealso{
Requires \code{\link[rgdal]{readOGR}},
  \code{\link[sp]{SpatialPoints}} and \code{\link[sp]{over}} to determine the
  country names.
}
