% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_lon180.R
\name{get_lon180}
\alias{get_lon180}
\title{Tests the Geographic Coordinates and Transforms the Longitude from [0, 360]
to [-180, +180].}
\usage{
## If there are the coordinates data frame
get_lon180(coords)
##
## If there are the coordinates text file
get_lon180(coords = NULL)
}
\arguments{
\item{coords}{data frame with the geographic coordinates in decimal degrees,
the longitude in the first column, the latitude in the second column and
the column names: \strong{lon | lat}.}
}
\description{
Given a list of geographic coordinates, first tests the coordinates for out
of bounds values and missing values. Then transforms the longitude from [0,
360] to [-180, +180], considering 6 significative digits.
}
\details{
\strong{Input:}
\itemize{
\item A text file without header with the geographic coordinates in decimal
degrees: the longitude in the first column and the latitude in the
second column, separated by tabs. Missing values must be codified as
'-999.0'.
\item Or a data frame with that same format and the column names: \strong{lon
| lat}.
}
\strong{Output:}
\itemize{
\item A text file named 'coords_lon180' with three columns \strong{id | lon |
lat} where: 'id' is the row identifier for the coordinates in the original
list of coordinates, 'lon' is the longitude in the range [-180, +180] and
'lat' is the latitude. Both coordinates are in decimal degrees.
\item If there are errors, the function writes a text file with the
coordinate pairs containing at least one coordinate out of bounds.
\item If there are missing coordinates, the function writes a text file with
the coordinate pairs containing at least one missing coordinate.
\item A .RData file  with the output data frame(s) which has/have the column
names: \strong{id | lon | lat}. The data frame 'coords_lon180' can be used as
input parameter of \code{\link{get_country}} or \code{\link{get_sea}} to
determine the country or the sea names. Eventually is created the data frame
'excl_coords' with the erroneous and/or missing coordinates.
}
}
\examples{
\dontrun{
##
## A sample of coordinate pairs with the longitude in the range [0, 360],
## extracted from the list of meteorological stations belonging to
## ISPD - International Surface Pressure Databank.
## Run the line below to see the example:
##
get_lon180(coords = coords_lon360)  # CRIAR O OBJECTO coords_lon360!!!!
}

}
