% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lon180.R
\name{get_lon180}
\alias{get_lon180}
\title{Transforms the longitude from [0, 360] to [-180, +180]}
\usage{
get_lon180()
}
\description{
Given a pair of geographic coordinates, transforms the longitude from [0,
360] to [-180, +180].
}
\details{
Input: a .txt file with the longitude in the first column and the latitude in
the second column, separated by tabs. The header of the file must be
'longitude' and 'latitude', respectively.

Opens a window, for choosing the input file that must be in the working
directory. Verify if the longitude and latitude are valid. Transforms the
longitude.

Output: creates a .txt file <coords2getf.txt> with the transformed longitude
and the latitude. This file is the input of \code{\link{get_country}} or
\code{\link{get_marine}} to determine the geographic names.
}
