% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_get_sea.R
\name{get_sea}
\alias{get_sea}
\title{Determines Sea Names for Points Located in the Sea.}
\usage{
get_sea()
}
\description{
Given a list of geographic coordinates in the sea, determines the Exclusive
Economic Zone (EEZ) name or the International Hydrographic Organization (IHO)
Sea name for each coordinate point.
}
\details{
\itemize{
\item The input is a .txt file with the geographic coordinates in decimal
degrees and the longitude in the range [-180, +180]. The longitude must be in
the first column and the latitude in the second column, separated by tabs.
The header of the file must be 'longitude' and 'latitude', respectively.
\item The function does the following:
\enumerate{
\item If exists 'countries_na.txt' created by \code{\link{get_country}},
automatically loads and reads the coordinates file;
\item Else if it is assumed that all the points are in the sea, opens a
window to choose the input file which must be formatted according to the
instructions above. In this case, tests if the longitude and latitude are out
of bounds and writes a file with the errors, if they exist;
\item Preserves or creates the 'id', necessary to display the coordinates in
the initial order at the end of the process;
\item Uses \code{\link[mregions]{mr_shp}} to access the Marineregions.org Web
Feature Service and get the shapefile with the polygons of the Exclusive
Economic Zones (EEZ) and the International Hydrographic Organization (IHO)
Seas, which covers most of marine regions on Earth;
\item Writes a file with points and their sea names and another file with
points that remain unnamed (if they exist). Also writes a report about the
points processed.
}
\item Output:
\enumerate{
\item A .txt file named 'err_coord-get_sea.txt' with the coordinates out of
bounds, if they exist;
\item A .txt file named 'seas.txt'. The header is: 'id' 'longitude' 'latitude'
'eez'/'eez_iho';
\item A .txt file named 'seas_na.txt' with the points unnamed, if they exist.
This file is the input of \code{\link{get_country_sea}} to determine the
remaining country or sea names;
\item A .txt file named 'report_get_sea.txt'.
}
}
}
\examples{
\dontrun{
##
## Execute one of the examples of \\code{\\link{get_country}} and then run the
## line below:
get_sea()
##
## About the sea names:
## Considering the Portuguese territory, if the point is in the Portuguese
## Continental Shelf, attributes the name 'Portuguese Exclusive Economic
## Zone'. If the point is not part of a continental shelf attributes the
## name of the sea /ocean (e.g. 'North Atlantic Ocean').
}

}
\references{
VLIZ (2012). Intersect of IHO Sea Areas and Exclusive Economic
  Zones (version 2). Available online at http://www.marineregions.org/.
  Consulted on 2017-11-15.
}
\seealso{
Requires \code{\link[mregions]{mr_shp}},
  \code{\link[sp]{SpatialPoints}}, \code{\link[sp]{CRS-class}} and
  \code{\link[sp]{over}} to determine the country names.
}
