% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_get_sea.R
\name{get_sea}
\alias{get_sea}
\title{Determines Sea Names for Points Located in a Marine Area.}
\usage{
get_sea(icoords)
}
\arguments{
\item{icoords}{data frame with three columns: \strong{id | lon | lat} where:
'id' is the row identifier for the coordinates in the original list of
coordinates, 'lon' is the longitude in the range [-180, +180] and 'lat' is
the latitude. Both coordinates are in decimal degrees.}
}
\description{
Given a list of geographic coordinates in marine areas of the world,
determines the geographical name of the sea, ocean, bay, gulf, fjord, etc.
were each one of the points is located. Uses a SpatialPolygonsDataFrame with
10 m precision containing the polygons and names of oceans/seas, which covers
most of the marine territories on Earth.
}
\details{
\strong{Input:}
\itemize{
\item The output of \code{\link{get_country}} or
\code{\link{get_country_shoreline}}, i.e., a data frame with three columns:
\strong{id | lon | lat}.
}
\strong{Output:}
\itemize{
\item Eventually, a text file with the points for which a sea name was
determined. That file has the following header: \strong{id | lon | lat |
sea}.
\item Eventually, a text file with the points missing sea names.
\item A .RData file with the sea names data frame - 'seas' -, the
missing sea names data frame - 'miss_seas' - or the both.
}
}
\examples{
\dontrun{
##
## First run
get_lon180(coords = ispd)
## Or
test_geocoord(coords = eraclim_uao_fp)
## Then run sequentially
get_country(icoords)
get_country_shoreline(icoords, tol)
get_sea(icoords)
}

}
\references{
Made with Natural Earth. Free vector and raster map data @
naturalearthdata.com.
}
