% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_order_data.R
\name{order_data}
\alias{order_data}
\title{Reorder the data by the coordinate pairs inicial order}
\usage{
order_data()
}
\description{
Concatenate all the files with country names and/or marine region names, i.
e, the output of \code{\link{get_country}}, \code{\link{get_marine}} and
\code{\link{get_country_sea}} and reorder the data by the coordinate pairs
inicial order.
}
