% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_seas_polys.R
\name{seas_polys}
\alias{seas_polys}
\title{Downloads Spatial Data with the World Seas Polygons and Saves it as a R
Object.}
\usage{
seas_polys()
}
\description{
Downloads from Natural Earth a shapefile with the world seas polygons of 10 m
precision and then transforms it into a SpatialPolygonsDataFrame which is
saved in the user's working directory. The data it contains is necessary for
the function \code{\link{get_sea}} to work.
}
\details{
Before run this function the user should verify that the working directory is
the correct one because the SpatialPolygonsDataFrame will be saved in the
working directory as a .RData file. That file is loaded from the working
directory when running the function \code{\link{get_sea}}.

\strong{Shapefile:}
\itemize{
\item http://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/physical/ne_10m_geography_marine_polys.zip
}
\strong{Shapefile metadata:}
\itemize{
\item http://www.naturalearthdata.com/downloads/10m-physical-vectors/10m-physical-labels/
}
}
\references{
Made with Natural Earth. Free vector and raster map data @
naturalearthdata.com.
}
